chore: Update .gitignore and Pipfile
This commit updates the .gitignore file to include the AWS config and credentials files. These files should not be tracked in the repository for security reasons.

It also adds the aws-cdk-lib and constructs dependencies to the Pipfile. These dependencies are required for the project and will be used for AWS CDK development.

These changes are necessary for proper project configuration and development with AWS CDK.